requirements.txt
Pillow==3.2.0
qiniu==7.0.7
requests==2.10.0
